Refactor routing fallback into router module (#21)
Move fallback behavior from main.js into router.js so route decisions are centralized.

- Add explicit login route and catch-all default route in router configuration

- Update route matcher to return configured route objects consistently

- Simplify main.js by removing manual unknown-route fallback branch

- Keep auth guard flow in main.js while delegating route matching to router.js
